+++ lib/design_system/nhsuk/builders/elements/backlink.rb
@@ -1,0 +1,41 @@
+# frozen_string_literal: true
+
+module DesignSystem
+  module Nhsuk
+    module Builders
+      module Elements
+        # This mixin module is used to provide back link.
+        module Backlink
+          private
+
+          def content_for_backlink
+            content_for(:backlink) do
+              content_tag(:nav) do
+                content_tag(:div, class: "#{brand}-width-container") do
+                  render_backlink
+                end
+              end
+            end
+          end
+
+          def render_backlink(label = @label, path = @path)
+            content_tag(:div, class: "#{brand}-back-link") do
+              content_tag(:a, class: "#{brand}-back-link__link", href: path) do
+                arrow_icon + label
+              end
+            end
+          end
+
+          # SVG for the arrow icon
+          def arrow_icon
+            %(
+            <svg class="#{brand}-icon #{brand}-icon__chevron-left" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true" height="24" width="24">
+            <path d="M8.5 12c0-.3.1-.5.3-.7l5-5c.4-.4 1-.4 1.4 0s.4 1 0 1.4L10.9 12l4.3 4.3c.4.4.4 1 0 1.4s-1 .4-1.4 0l-5-5c-.2-.2-.3-.4-.3-.7z"></path>
+            </svg>
+            ).html_safe
+          end
+        end
+      end
+    end
+  end
+end