+++ lib/design_system/nhsuk/builders/fixed_elements.rb
@@ -5,6 +5,7 @@ module DesignSystem
     module Builders
       # This class is used to provide the NHSUK fixed elements builder.
       class FixedElements < ::DesignSystem::Govuk::Builders::FixedElements
+        include Elements::Backlink
         include Elements::Breadcrumbs
       end
     end