+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `back_link` to fixed elements 
+
 ## [0.11.0] - 2026-01-16
 
 ### Changed